requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
pi==0.1.2
python-dotenv==1.0.0
sympy==1.11.1
Werkzeug==2.3.0
